winewayland: Use system cursor shapes when possible.
Use the cursor-shapes-v1 protocol to tell the compositor which system
cursor shape to use. If a shape match is not found (or cursor-shapes-v1
is not available) we fall back to setting the cursor buffer from the
Windows cursor data as before.

Wine-Bug: https://bugs.winehq.org/show_bug.cgi?id=57648